hot fix

+++ 0x10-variadic_functions/3-print_all.c
@@ -32,8 +32,8 @@ void print_int(va_list arg)
 {
 	int num;
 
-	num = va_arg(arg, double);
-	printf("%f", num);
+	num = va_arg(arg, int);
+	printf("%d", num);
 }
 /**
  * print_float - Prints a float.
